Refactor screen health integration to use separate preset files
